fix: prevent i32 overflow in score/target computations and use strip_prefix for string slicing safety
Widen bestscore*notes/totalnotes and rivalscore*notes/totalnotes to i64
before dividing in score_data_property to prevent overflow on high
note-count charts (33k+ notes). Apply the same i64 widening to
total*value/100 in target_property IR rank-rate computations.

Replace byte-index slicing (id[8..], id[12..]) with strip_prefix in
InternetRankingTargetProperty::from_id to match the safer pattern
already used by RIVAL_* variants.

+++ crates/rubato-core/src/score_data_property.rs
@@ -197,7 +197,7 @@ impl ScoreDataProperty {
             self.nowbestscore = if totalnotes == 0 {
                 0
             } else {
-                self.bestscore * notes / totalnotes
+                (self.bestscore as i64 * notes as i64 / totalnotes as i64) as i32
             };
             self.nowbestscorerate = if totalnotes == 0 {
                 0.0
@@ -222,7 +222,7 @@ impl ScoreDataProperty {
             self.nowrivalscore = if totalnotes == 0 {
                 0
             } else {
-                self.rivalscore * notes / totalnotes
+                (self.rivalscore as i64 * notes as i64 / totalnotes as i64) as i32
             };
             self.nowrivalscorerate = if totalnotes == 0 {
                 0.0
